requirements.txt
openai>=1.98.0
playwright>=1.52.0
python-dotenv>=1.0.1
Flask>=3.0.0
gunicorn>=23.0.0
